// src/app/core/models/auth/user.model.ts
export class UserModel {  
    username: string;
    password: string;
    token: string;
    _id: string;
    role: string;
    designationid: string;
    branchid: string;
    roletype: string;
    currency: string;
    language: string;
    cloud_name: string;
    user: any;
    property: any;
    rtl: boolean;
    fullname: string;
    namenumber: string;
    profileimage: string;
    status: string;
}